Made new function for creating variables
This function makes it easy to specify the parameters of a variable’s
distribution, and how it should be coded. It creates a DGP_object that
is used by both make_covariates() and make_potential_outcomes() to
specify how the data is generated (i.e. if some covariates multinomial,
others linear, outcome is binomial, etc.).